+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/savings/api/SavingsProductsApiResourceSwagger.java
@@ -40,7 +40,7 @@ final class SavingsProductsApiResourceSwagger {
             private PostSavingsCharges() {}
 
             @Schema(example = "1")
-            public Integer id;
+            public Long id;
         }
 
         @Schema(example = "Passbook Savings")
@@ -80,7 +80,7 @@ final class SavingsProductsApiResourceSwagger {
         private PostSavingsProductsResponse() {}
 
         @Schema(example = "1")
-        public Integer resourceId;
+        public Long resourceId;
     }
 
     @Schema(description = "PutSavingsProductsProductIdRequest")
@@ -114,7 +114,7 @@ final class SavingsProductsApiResourceSwagger {
         }
 
         @Schema(example = "1")
-        public Integer resourceId;
+        public Long resourceId;
         public PutSavingsChanges changes;
     }
 
@@ -231,7 +231,7 @@ final class SavingsProductsApiResourceSwagger {
             private GetSavingsProductsGlAccount() {}
 
             @Schema(example = "12")
-            public Integer id;
+            public Long id;
             @Schema(example = "savings control")
             public String name;
             @Schema(example = "2000001")
@@ -265,7 +265,7 @@ final class SavingsProductsApiResourceSwagger {
                 private GetSavingsProductsPaymentType() {}
 
                 @Schema(example = "10")
-                public Integer id;
+                public Long id;
                 @Schema(example = "check")
                 public String name;
             }
@@ -275,7 +275,7 @@ final class SavingsProductsApiResourceSwagger {
                 private GetSavingsProductsFundSourceAccount() {}
 
                 @Schema(example = "12")
-                public Integer id;
+                public Long id;
                 @Schema(example = "savings ref")
                 public String name;
                 @Schema(example = "20")
@@ -295,7 +295,7 @@ final class SavingsProductsApiResourceSwagger {
                 private GetSavingsProductsFeeToIncomeAccountMappingsCharge() {}
 
                 @Schema(example = "11")
-                public Integer id;
+                public Long id;
                 @Schema(example = "sav charge")
                 public String name;
                 @Schema(example = "false")
@@ -309,7 +309,7 @@ final class SavingsProductsApiResourceSwagger {
                 private GetSavingsProductsFeeToIncomeAccountMappingsIncomeAccount() {}
 
                 @Schema(example = "16")
-                public Integer id;
+                public Long id;
                 @Schema(example = "income from savings fee")
                 public String name;
                 @Schema(example = "24")
@@ -329,7 +329,7 @@ final class SavingsProductsApiResourceSwagger {
                 private GetSavingsProductsPenaltyToIncomeAccountMappingsCharge() {}
 
                 @Schema(example = "12")
-                public Integer id;
+                public Long id;
                 @Schema(example = "sav 2")
                 public String name;
                 @Schema(example = "false")
@@ -343,7 +343,7 @@ final class SavingsProductsApiResourceSwagger {
         }
 
         @Schema(example = "1")
-        public Integer id;
+        public Long id;
         @Schema(example = "savings product")
         public String name;
         @Schema(example = "sa1")
@@ -365,7 +365,7 @@ final class SavingsProductsApiResourceSwagger {
         public Set<GetSavingsProductsFeeToIncomeAccountMappings> feeToIncomeAccountMappings;
         public Set<GetSavingsProductsPenaltyToIncomeAccountMappings> penaltyToIncomeAccountMappings;
         @Schema(example = "[]")
-        public List<Integer> charges;
+        public List<Long> charges;
     }
 
     @Schema(description = "GetSavingsProductsTemplateResponse")
@@ -462,11 +462,11 @@ final class SavingsProductsApiResourceSwagger {
                     private GetSavingsProductsLiabilityTagId() {}
 
                     @Schema(example = "0")
-                    public Integer id;
+                    public Long id;
                 }
 
                 @Schema(example = "15")
-                public Integer id;
+                public Long id;
                 @Schema(example = "Savings Control")
                 public String name;
                 @Schema(example = "50001")
@@ -504,7 +504,7 @@ final class SavingsProductsApiResourceSwagger {
                 }
 
                 @Schema(example = "2")
-                public Integer id;
+                public Long id;
                 @Schema(example = "Cash")
                 public String name;
                 @Schema(example = "100001")
@@ -535,7 +535,7 @@ final class SavingsProductsApiResourceSwagger {
                 }
 
                 @Schema(example = "6")
-                public Integer id;
+                public Long id;
                 @Schema(example = "Write Off Expenses")
                 public String name;
                 @Schema(example = "60001")
@@ -566,7 +566,7 @@ final class SavingsProductsApiResourceSwagger {
                 }
 
                 @Schema(example = "3")
-                public Integer id;
+                public Long id;
                 @Schema(example = "income from interest")
                 public String name;
                 @Schema(example = "40001")
@@ -649,7 +649,7 @@ final class SavingsProductsApiResourceSwagger {
             public GetSavingsChargeTimeType chargeTimeType;
             public GetSavingsProductsResponse.GetSavingsCurrency currency;
             @Schema(example = "4")
-            public Integer id;
+            public Long id;
             @Schema(example = "Savings charge 1")
             public String name;
             @Schema(example = "false")
@@ -683,6 +683,6 @@ final class SavingsProductsApiResourceSwagger {
         private DeleteSavingsProductsProductIdResponse() {}
 
         @Schema(example = "1")
-        public Integer resourceId;
+        public Long resourceId;
     }
 }
